+++ src/main/java/com/LetMeDoWith/LetMeDoWith/domain/task/repository/DowithTaskQueryRepository.java
@@ -3,7 +3,6 @@ package com.LetMeDoWith.LetMeDoWith.domain.task.repository;
 import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.DowithTaskDetailQueryDto;
 import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.DowithTaskQueryDto;
 import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.SuccessDowithTaskQueryDto;
-
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -15,8 +14,6 @@ public interface DowithTaskQueryRepository {
     List<DowithTaskQueryDto> getDowithTasks(String memberId, LocalDate startDate, LocalDate endDate);
 
     Optional<DowithTaskDetailQueryDto> getDowithTask(String memberId, Long dowithTaskId);
-
-    Long countSuccessDowithTasks();
 
     List<SuccessDowithTaskQueryDto> getSuccessDowithTasks(String requestMemberId, int offset, int limit);
