+++ src/main/java/com/LetMeDoWith/LetMeDoWith/domain/task/repository/dto/SuccessDowithTaskQueryDto.java
@@ -1,5 +1,4 @@
 package com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto;
 
 public record SuccessDowithTaskQueryDto(
-        Long id, String title, String nickname, String profileImageUrl, String successImageUrl, Boolean isLiked) {
-}
+        Long id, String title, String nickname, String profileImageUrl, String successImageUrl, Boolean isLiked) {}
